% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_steps_interface.R
\name{All_steps_interface}
\alias{All_steps_interface}
\title{A function for conducting the algorithm through the toy data}
\usage{
All_steps_interface(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  gene_exp = gene_exp,
  mirna_exp = mirna_exp,
  surv_data = surv_data,
  filtering = "less",
  window_size = 10,
  cor_method = "pearson",
  cor_threshold_peak = 0.85
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default: demo)}

\item{disease_name}{the abbreviation of disease that users are interested in
(default: DLBC)}

\item{gene_exp}{location of gene expression data (default: gene_exp)}

\item{mirna_exp}{location of miRNA expression data (default: mirna_exp)}

\item{surv_data}{location of survival data (default: surv_data)}

\item{filtering}{three different filtering criteria, including strict,
moderate and less. (Default: less)}

\item{window_size}{the number of samples for each window (default:10)}

\item{cor_method}{selection of correlation methods, including pearson and
spearman (default: pearson)}

\item{cor_threshold_peak}{peak threshold of correlation value between 0 and 1
(default: 0.85)}
}
\value{
a dataframe object
}
\description{
A function to allow users to conducting the
}
\examples{
data(gene_exp)
data(mirna_exp)
data(surv_data)
All_steps_interface(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC',
gene_exp = gene_exp,
mirna_exp = mirna_exp,
surv_data = surv_data,
filtering = 'less',
window_size = 10,
cor_method = 'pearson',
cor_threshold_peak = 0.85)
}
