% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAFunction.R
\name{ceRNAFunction}
\alias{ceRNAFunction}
\title{Functional enrichment analysis}
\usage{
ceRNAFunction(path_prefix = NULL, project_name, disease_name, pairs_cutoff)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{window_size}{the number of samples for each window and usually about
one third of total samples}

\item{cor_method}{selection of correlation methods, including pearson and
spearman (default: pearson)}

\item{cor_threshold_peak}{peak threshold of correlation value between 0 and 1
(default: 0.85)}
}
\description{
A function to conduct Functional enrichment analysis for
biological interpretation. The databases supported in this function include
Gene Ontology (GO; http://geneontology.org/) and Kyoto Encyclopedia of
Genes and Genomes (KEGG; https://www.genome.jp/kegg/).
}
\examples{
load('data/ExampleData.Rdata')
ceRNAFunction(
project_name = 'demo',
disease_name = 'DLBC',
pairs_cutoff = 1
)

}
