% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNALocation.R
\name{ceRNALocation}
\alias{ceRNALocation}
\title{Visualization for peak location}
\usage{
ceRNALocation(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  mirna = "hsa-miR-101-3p",
  window_size = 10
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{mirna}{a specific mirna name (such as hsa-miR-101-3p) assigned by user}

\item{window_size}{the number of samples for each window and usually about
one third of total samples}
}
\value{
a list of figure object
}
\description{
A function to visualize the peak location at certain miRNA level
}
\examples{
ceRNALocation(
project_name = 'demo',
disease_name = 'DLBC',
mirna='hsa-miR-101-3p',
window_size = 10
)


}
