% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAMethod.R
\name{ceRNAMethod}
\alias{ceRNAMethod}
\title{Main ceRNAR algorithm}
\usage{
ceRNAMethod(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  window_size = 10,
  cor_method = "pearson",
  cor_threshold_peak = 0.85
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default: demo)}

\item{disease_name}{the abbreviation of disease that users are interested in
(default: DLBC)}

\item{window_size}{the number of samples for each window (defaut: 10)}

\item{cor_method}{selection of correlation methods, including pearson and
spearman (default: pearson)}

\item{cor_threshold_peak}{peak threshold of correlation value between 0 and 1
(default: 0.85)}
}
\value{
a dataframe object
}
\description{
A function to conduct three steps in algorithm, including pairs
filtering, segment clustering and peak merging
}
\examples{
ceRNAMethod(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC',
window_size = 10,
cor_method = 'pearson',
cor_threshold_peak = 0.85
)


}
