% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAModule.R
\name{ceRNAModule}
\alias{ceRNAModule}
\title{Network analysis and visualization}
\usage{
ceRNAModule(
  path_prefix,
  project_name = "demo",
  disease_name = "DLBC",
  pairs_cutoff = 5,
  column_sum = 1
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{pairs_cutoff}{at least the number of ceRNA pairs that a mirna must have}

\item{column_sum}{the number of ceRNAs}
}
\value{
a list of figure object
}
\description{
A function to analyze and visualize potential network of
identified ceRNAs
}
\examples{
ceRNAModule(
path_prefix = '~/',
project_name = 'demo',
disease_name = 'DLBC',
pairs_cutoff = 5,
column_sum = 1
)


}
