% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNApairFiltering.R
\name{ceRNApairFiltering}
\alias{ceRNApairFiltering}
\alias{ceRNApairFilering}
\title{one of three steps in main ceRNAR algorithm}
\usage{
ceRNApairFilering(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  window_size = 10,
  cor_method = "pearson"
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default: demo)}

\item{disease_name}{the abbreviation of disease that users are interested in
(default: DLBC)}

\item{window_size}{the number of samples for each window (default: 10)}

\item{cor_method}{selection of correlation methods, including pearson and
spearman (default: pearson)}
}
\value{
file
}
\description{
A function to conduct one of three steps in algorithm,
that is, pairs filtering
}
\examples{
ceRNApairFilering(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC',
window_size = 10,
cor_method = 'pearson'
)


}
