% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAputativePairs.R
\name{ceRNAputativePairs}
\alias{ceRNAputativePairs}
\title{Extraction of putative mRNA-miRNA pairs}
\usage{
ceRNAputativePairs(
  path_prefix,
  project_name = "demo",
  disease_name = "DLBC",
  filtering = "less"
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{filtering}{three different filtering criteria, including strict,
moderate and less. (Default: less)}
}
\value{
a file
}
\description{
A function to obtain putative mRNA-miRNA pairs from several databases
}
\examples{
ceRNAputativePairs(
path_prefix = '~/',
project_name ='demo',
disease_name = 'DLBC',
filtering = 'less'
)


}
