% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNASurvival.R
\name{ceRNASurvival}
\alias{ceRNASurvival}
\title{Survival analysis and visualization}
\usage{
ceRNASurvival(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  mirnas = "hsa-miR-101-3p"
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{mirnas}{a list of mirna name}
}
\value{
a list object
}
\description{
A function to analyze the survival outcome when people carry the
identified ceRNAs and visualize the results using Kaplan-Meier plot
}
\examples{
ceRNASurvival(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC',
mirnas = 'hsa-miR-101-3p'
)


}
