% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAValidate.R
\name{ceRNAValidate}
\alias{ceRNAValidate}
\title{Externally experimental validation for the potential ceRNA pairs}
\usage{
ceRNAValidate(path_prefix = NULL, project_name = "demo", disease_name = "DLBC")
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}
}
\value{
a dataframe object
}
\description{
A function to validate the potential ceRNA pairs based on the miRSponge database
(http://www.bio-bigdata.net/miRSponge)
}
\examples{
ceRNAValidate(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC'
)


}
