% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp_data.R
\docType{data}
\name{gene_exp}
\alias{gene_exp}
\title{Example gene expression dataset for ceRNAR}
\format{

}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(gene_exp)
}
\description{
A data set with the gene expression for 45 subjects
}
\examples{
data(gene_exp)
}
\keyword{datasets}
