% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirna_exp_data.R
\docType{data}
\name{mirna_exp}
\alias{mirna_exp}
\title{Example mirna expression dataset for ceRNAR}
\format{

}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(mirna_exp)
}
\description{
A data set with the mirna expression for 45 subjects
}
\examples{
data(mirna_exp)
}
\keyword{datasets}
