% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentClusteringPlusPeakMerging.R
\name{SegmentClusteringPlusPeakMerging}
\alias{SegmentClusteringPlusPeakMerging}
\title{two of three steps in main ceRNAR algorithm}
\usage{
SegmentClusteringPlusPeakMerging(
  path_prefix = NULL,
  project_name,
  disease_name,
  cor_threshold_peak = 0.85,
  window_size
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{cor_threshold_peak}{peak threshold of correlation value between 0 and 1 (default: 0.85)}

\item{window_size}{the number of samples for each window}
}
\description{
A function to conduct two of three steps in algorithm, that is, segment clustering and peak merging
}
\examples{
SegmentClusteringPlusPeakMerging(
project_name = 'demo',
disease_name="DLBC",
window_size = 45/5
)

}
