% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_data.R
\docType{data}
\name{surv_data}
\alias{surv_data}
\title{Example survival dataset for ceRNAR}
\format{

}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(surv_data)
}
\description{
A data set with the survival outcome for 45 subjects
}
\examples{
data(surv_data)
}
\keyword{datasets}
