% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNACustomize.R
\name{ceRNACustomize}
\alias{ceRNACustomize}
\title{A function for uploading customized data}
\usage{
ceRNACustomize(
  path_prefix = NULL,
  project_name = NULL,
  disease_name = NULL,
  gene_exp = NULL,
  mirna_exp = NULL,
  surv_data
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign}

\item{disease_name}{the abbreviation of disease that users are interested in}

\item{gene_exp}{location of gene expression data}

\item{mirna_exp}{location of miRNA expression data}

\item{surv_data}{location of survival data}
}
\description{
A function to allow users to upload their own data
}
\examples{
ceRNACustomize(
project_name = 'demo',
disease_name = 'DLBC',
gene_exp = gene_exp,
mirna_exp = mirna_exp,
surv_data = surv_data
)

}
