% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAFunction.R
\name{ceRNAFunction}
\alias{ceRNAFunction}
\title{Functional enrichment analysis}
\usage{
ceRNAFunction(
  path_prefix = NULL,
  project_name = "demo",
  disease_name = "DLBC",
  pairs_cutoff = 1,
  kegg_p_cutoff = 0.05,
  kegg_top = 10,
  go_p_cutoff = 0.05,
  go_q_cutoff = 0.05,
  go_top = 10
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default: demo)}

\item{disease_name}{the abbreviation of disease that users are interested in
(default: DLBC)}

\item{pairs_cutoff}{at least the number of ceRNA pairs that a mirna must have
(default: 1)}

\item{kegg_p_cutoff}{a predifined cutoff value of p-value from enrichment analysis based on KEGG database (default: 0.01)}

\item{kegg_top}{the number of top KEGG category to show in the graphic display (default: 10)}

\item{go_p_cutoff}{a predifined cutoff value of p-value from enrichment analysis based on GO database (default: 0.05)}

\item{go_q_cutoff}{a predifined cutoff value of p-value from enrichment analysis based on GO database (default: 0.05)}

\item{go_top}{the number of top GO category to show in the graphic display (default: 10)}
}
\value{
a list object
}
\description{
A function to conduct Functional enrichment analysis for
biological interpretation. The databases supported in this function include
Gene Ontology (GO; http://geneontology.org/) and Kyoto Encyclopedia of
Genes and Genomes (KEGG; https://www.genome.jp/kegg/).
}
\examples{
ceRNAFunction(
path_prefix = NULL,
project_name = 'demo',
disease_name = 'DLBC',
pairs_cutoff = 1,
kegg_p_cutoff = 0.05,
kegg_top = 10,
go_p_cutoff = 0.05,
go_q_cutoff = 0.05,
go_top = 10
)


}
