% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAFunction.R
\name{ceRNAFunction}
\alias{ceRNAFunction}
\title{Functional enrichment analysis}
\usage{
ceRNAFunction(
  path_prefix,
  project_name = "demo",
  disease_name = "DLBC",
  pairs_cutoff = 1
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default: demo)}

\item{disease_name}{the abbreviation of disease that users are interested in
(default: DLBC)}

\item{pairs_cutoff}{at least the number of ceRNA pairs that a mirna must have
(default: 1)}
}
\value{
a list of figure objects
}
\description{
A function to conduct Functional enrichment analysis for
biological interpretation. The databases supported in this function include
Gene Ontology (GO; http://geneontology.org/) and Kyoto Encyclopedia of
Genes and Genomes (KEGG; https://www.genome.jp/kegg/).
}
\examples{
ceRNAFunction(
path_prefix = '~/',
project_name = 'demo',
disease_name = 'DLBC',
pairs_cutoff = 1
)


}
