% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNATCGA.R
\name{ceRNATCGA}
\alias{ceRNATCGA}
\title{Retrieval of public TCGA data from GDC Xena Hub}
\usage{
ceRNATCGA(
  path_prefix = NULL,
  project_name = "TCGA",
  disease_name = "DLBC",
  timeout = 5e+05
)
}
\arguments{
\item{path_prefix}{user's working directory}

\item{project_name}{the project name that users can assign (default = 'TCGA')}

\item{disease_name}{the abbreviation of disease that users are interested in
(default = 'DLBC')}

\item{timeout}{the allowance time for downloading TCGA data
(default = 1000)}
}
\value{
file
}
\description{
A function to retrieve TCGA data from GDC Xena Hub
(https://xenabrowser.net/datapages/)
}
\examples{
ceRNATCGA(
project_name = 'TCGA',
disease_name = 'DLBC',
timeout = 500000
)

}
