% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp_data.R
\docType{data}
\name{gene_exp}
\alias{gene_exp}
\title{Example gene expression dataset for ceRNAR}
\format{
A R data frame with 185 rows and 45 variables
\describe{
  \item{expression data}{Numeric gene expression matrix with each columns
   as a different individual and each row as a separate gene.}
}
}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(gene_exp)
}
\description{
A data set with the gene expression for 45 subjects
}
\examples{
data(gene_exp)
}
\keyword{datasets}
