% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirna_exp_data.R
\docType{data}
\name{mirna_exp}
\alias{mirna_exp}
\title{Example mirna expression dataset for ceRNAR}
\format{
A R data frame with 78 rows and 45 variables
\describe{
  \item{expression data}{Numeric mirna expression matrix with each
  columns as a different individual and each row as a separate miRNA.}}
}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(mirna_exp)
}
\description{
A data set with the mirna expression for 45 subjects
}
\examples{
data(mirna_exp)
}
\keyword{datasets}
