% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_data.R
\docType{data}
\name{surv_data}
\alias{surv_data}
\title{Example survival dataset for ceRNAR}
\format{
A R data frame with 78 rows and 45 variables
\describe{
  \item{survival data}{Numeric survival matrix with two columns,
  including status and time, and each row as a different individual.}
}
}
\source{
<https://www.github.com/ywhsiao/ceRNAR>
}
\usage{
data(surv_data)
}
\description{
A data set with the survival outcome for 45 subjects
}
\examples{
data(surv_data)
}
\keyword{datasets}
